requirements.txt
langchain
langchain-groq
groq
streamlit
pandas
python-dotenv
tabulate